ai-document-generator/requirements.txt
streamlit>=1.33
python-dotenv>=1.0
openai>=1.30
